[UPD] Create a models and views subdirectories and split the source code by objects
