+++ stock_location_inactive/models/__init__.py
@@ -21,6 +21,6 @@
 ##############################################################################
 
 from . import (
-    models,
-    report,
+    stock_location,
+    stock_move,
 )
