+++ stock_location_inactive/models/stock_move.py
@@ -20,7 +20,35 @@
 #
 ##############################################################################
 
-from . import (
-    models,
-    report,
-)
+from openerp.osv import orm
+from openerp.tools.translate import _
+
+
+class StockMove(orm.Model):
+
+    """
+    Subclass stock.move model.
+
+    Extend the model stock.move to add a constraint to check if
+    the location is disable
+    """
+    _name = "stock.move"
+    _inherit = "stock.move"
+
+    def action_done(self, cr, uid, ids, context=None):
+        """Override the action_done to add constraint to check if the
+        location_dest is disable"""
+
+        if context is None:
+            context = {}
+        if isinstance(ids, (int, long)):
+            ids = [ids]
+        active = self.browse(cr, uid, ids, context)[0].location_dest_id.active
+        if not active:
+            raise orm.except_orm(
+                _('Validation Error'),
+                _('You cannot disable a location that contains products '
+                  'or sub-locations')
+            )
+        return super(StockMove, self).action_done(
+            cr, uid, ids, context=context)
