+++ stock_location_inactive/views/stock_view.xml
